pmoves/docs/PMOVES.AI PLANS/Enhanced Media Stack with Advanced AudioVideo Analysis/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
librosa==0.10.1
pydub==0.25.1
numpy>=1.24.0
scipy>=1.10.0
soundfile>=0.12.0
python-multipart==0.0.6
pydantic>=2.0.0
whisper-openai
pyannote.audio
speechbrain
datasets
accelerate
